refactor: ♻️ refactor datatables to remove some code duplication
more work needed
