+++ app/src/components/students/list/columns.tsx
@@ -4,7 +4,7 @@ import { Skeleton } from '@pedaki/design/ui/skeleton';
 import type { Field } from '@pedaki/services/students/query.model.client';
 import type { GetManyStudentsOutput } from '@pedaki/services/students/student.model.js';
 import type { ColumnDef } from '@tanstack/react-table';
-import { DataTableColumnHeader } from '~/components/students/list/data-table-column-header.tsx';
+import { DataTableColumnHeader } from '~/components/datatable/data-table-column-header.tsx';
 import type { UseScopedI18nType } from '~/locales/client.ts';
 import type { OutputType } from '~api/router/router.ts';
