+++ app/src/locales/fr/components/index.ts
@@ -1,0 +1,5 @@
+import datatable from '~/locales/fr/components/datatable';
+
+export default {
+  datatable: datatable,
+} as const;
